[Issue #1] Update template and projecct structure [ThinhLP]
