+++ .idea/artifacts/CocShopSystemVer2_war_exploded.xml
@@ -13,7 +13,6 @@
           <element id="library" level="project" name="Spring-4.3.9.RELEASE" />
           <element id="library" level="project" name="Spring MVC-4.3.9.RELEASE" />
           <element id="library" level="project" name="Spring Data JPA-1.0.0" />
-          <element id="library" level="project" name="mysql-connector-java-5.1.40" />
           <element id="library" level="project" name="jackson-core-2.8.8" />
           <element id="library" level="project" name="slf4j-api-1.7.25" />
         </element>
