--- .idea/libraries/mysql_connector_java_5_1_40.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="mysql-connector-java-5.1.40">
-    <CLASSES>
-      <root url="jar://C:/Program Files/MySQL/mysql-connector-java-5.1.40.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
