+++ .idea/libraries/mysql_connector_java_5_1_41_bin.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="mysql-connector-java-5.1.41-bin">
+    <CLASSES>
+      <root url="jar://C:/Program Files/MySQL/mysql-connector-java-5.1.41/mysql-connector-java-5.1.41-bin.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
