+++ CocShopSystemVer2.iml
@@ -47,8 +47,8 @@
     <orderEntry type="library" name="Spring-4.3.9.RELEASE" level="project" />
     <orderEntry type="library" name="Spring MVC-4.3.9.RELEASE" level="project" />
     <orderEntry type="library" name="Spring Data JPA-1.0.0" level="project" />
-    <orderEntry type="library" name="mysql-connector-java-5.1.40" level="project" />
     <orderEntry type="library" name="jackson-core-2.8.8" level="project" />
     <orderEntry type="library" name="slf4j-api-1.7.25" level="project" />
+    <orderEntry type="library" name="mysql-connector-java-5.1.41-bin" level="project" />
   </component>
 </module>
