+++ web/WEB-INF/applicationContext.xml
@@ -11,7 +11,7 @@
         <property name="driverClassName" value="com.mysql.jdbc.Driver"/>
         <property name="url" value="jdbc:mysql://localhost:3306/cocshop"/>
         <property name="username" value="root"/>
-        <property name="password" value="JokerZ0511"/>
+        <property name="password" value="123456"/>
     </bean>
     <jpa:repositories base-package="com.cocshop.repository"/>
 </beans>
